--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "adaptor"]
-	path = adaptor
-	url = git@github.com:antlr/jetbrains.git
