+++ gradle.properties
@@ -1,0 +1,8 @@
+# Available idea versions:
+# https://www.jetbrains.com/intellij-repository/releases
+# https://www.jetbrains.com/intellij-repository/snapshots
+
+version=0.1
+ideaVersion=145.256.33
+javaVersion=1.7
+buildNumber=SNAPSHOT
