--- src/java/org/antlr/jetbrains/sample/psi/FunctionSubtree.java
@@ -1,30 +1,0 @@
-package org.antlr.jetbrains.sample.psi;
-
-import com.intellij.lang.ASTNode;
-import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiNamedElement;
-import org.antlr.jetbrains.adaptor.SymtabUtils;
-import org.antlr.jetbrains.adaptor.psi.IdentifierDefSubtree;
-import org.antlr.jetbrains.adaptor.psi.ScopeNode;
-import org.antlr.jetbrains.sample.SampleLanguage;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-/** A subtree associated with a function definition.
- *  Its scope is the set of arguments.
- */
-public class FunctionSubtree extends IdentifierDefSubtree implements ScopeNode {
-	public FunctionSubtree(@NotNull ASTNode node) {
-		super(node);
-	}
-
-	@Nullable
-	@Override
-	public PsiElement resolve(PsiNamedElement element) {
-//		System.out.println(getClass().getSimpleName()+
-//			                   ".resolve("+myElement.getName()+
-//			                   " at "+Integer.toHexString(myElement.hashCode())+")");
-		return SymtabUtils.resolve(this, SampleLanguage.INSTANCE,
-		                           element, "/script/function/ID");
-	}
-}
