--- src/java/org/antlr/jetbrains/sample/psi/VardefSubtree.java
@@ -1,11 +1,0 @@
-package org.antlr.jetbrains.sample.psi;
-
-import com.intellij.lang.ASTNode;
-import org.antlr.jetbrains.adaptor.psi.IdentifierDefSubtree;
-import org.jetbrains.annotations.NotNull;
-
-public class VardefSubtree extends IdentifierDefSubtree {
-	public VardefSubtree(@NotNull ASTNode node) {
-		super(node);
-	}
-}
