+++ src/main/java/org/antlr/jetbrains/sample/SampleSyntaxHighlighter.java
