+++ src/main/java/org/antlr/jetbrains/sample/SampleSyntaxHighlighterFactory.java
