+++ src/main/java/org/antlr/jetbrains/sample/structview/SampleStructureViewElement.java
