+++ src/main/java/org/antlr/jetbrains/sample/structview/SampleStructureViewFactory.java
