+++ src/main/java/org/antlr/jetbrains/sample/structview/SampleStructureViewModel.java
