+++ src/main/resources/icons/org/antlr/jetbrains/sample/sample.png
